Restyled Groups page and added member list functionality
