+++ react_source/app/summary.js
@@ -1,12 +1,12 @@
 import * as globals from '../utils/globals.js'
 
 import Base from '../components/Base.js';
-import Groups from '../components/Groups.js';
+import GroupsList from '../components/GroupsList.js';
 
 export default function Page() {
     return (
         <Base style={globals.styles.container}>
-            <Groups />
+            <GroupsList />
         </Base>
     );
 }
